[chore] #6 - SwiftLint 경고 관련 코드 수정 및 리뷰  반영
